+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 176:
 
 * Tidy up Quick Reference
 * Fix array end calculation in utf8 assertion
+* WebSocket masks use secure PRNG by default
 
 --------------------------------------------------------------------------------